% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_projected_rank_odds.R
\name{generate_projected_rank_odds}
\alias{generate_projected_rank_odds}
\title{Generate Projected Rank Odds Table}
\usage{
generate_projected_rank_odds(projection_outputs)
}
\arguments{
\item{projection_outputs}{A list containing the outputs from the \code{run_projection} function, including mean projection odds.}
}
\value{
A gt table displaying the projected rank odds for each team.
}
\description{
This function generates a table displaying the projected rank odds for each team using the output from the season projection simulations.
}
\details{
This function creates a visually styled table showing the projected rank odds for each team based on the simulation results. It includes the probability of each team finishing in each rank position, formatted as percentages.
}
\examples{
\dontrun{
  projection_outputs <- run_projection(trials = 100, matches, deductions, date = Sys.Date(),
                                       league_adj = TRUE, xG_factor = 0.6, smooth = TRUE, hot = TRUE,
                                       europe = 6, rel = 18)
  rank_odds_table <- generate_projected_rank_odds(projection_outputs)
  print(rank_odds_table)
}

}
