% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_odds.R
\name{get_match_odds}
\alias{get_match_odds}
\title{Calculate Match Odds from Simulated Results}
\usage{
get_match_odds(simmed_results, trials)
}
\arguments{
\item{simmed_results}{A data frame containing the simulated results from multiple trials, including columns for game ID, date, game week (GW), home team, away team, home win, draw, and away win.}

\item{trials}{An integer specifying the number of trials run.}
}
\value{
A data frame with the calculated odds for each match, including columns for game ID, date, game week (GW), home team, away team, home win odds, draw odds, and away win odds.
}
\description{
This function calculates the odds for each match based on the simulated results from multiple trials.
}
\details{
This function calculates the probability of each match outcome (home win, draw, away win) based on the simulation results. It then converts these probabilities into odds.
}
\examples{
\dontrun{
  simmed_results <- data.frame(gameID = rep(1:10, each = 100),
                               Date = rep(Sys.Date(), 1000),
                               GW = rep(1:10, each = 100),
                               home_team = sample(letters[1:20], 1000, replace = TRUE),
                               away_team = sample(letters[1:20], 1000, replace = TRUE),
                               home_win = sample(0:1, 1000, replace = TRUE),
                               draw = sample(0:1, 1000, replace = TRUE),
                               away_win = sample(0:1, 1000, replace = TRUE))
  trials <- 100
  match_odds <- get_match_odds(simmed_results, trials)
}

}
