% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches_to_results.R
\name{matches_to_results}
\alias{matches_to_results}
\title{convert list of matches to list of results to be played}
\usage{
matches_to_results(matches, date = Sys.Date())
}
\arguments{
\item{date}{}
}
\description{
convert list of matches to list of results to be played
}
