% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_projection.R
\name{run_projection}
\alias{run_projection}
\title{run projection for given matches}
\usage{
run_projection(
  trials,
  matches,
  date = sys.Date(),
  league_adj,
  xG_factor,
  smooth,
  hot,
  europe = 6,
  rel = 18
)
}
\arguments{
\item{rel}{}
}
\description{
run projection for given matches
}
