% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_season.R
\name{simulate_season}
\alias{simulate_season}
\title{Title}
\usage{
simulate_season(
  fixtures,
  results,
  team_ratings,
  league_adj,
  xG_factor,
  smooth,
  hot
)
}
\arguments{
\item{results}{}
}
\description{
Title
}
