% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_matches.R
\name{split_matches}
\alias{split_matches}
\title{splits matches table to results, standings and remaining fixtures as at date}
\usage{
split_matches(matches, date = Sys.Date(), xG_factor, xTable = F)
}
\arguments{
\item{xG_factor}{}
}
\description{
splits matches table to results, standings and remaining fixtures as at date
}
