% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_final_projection.R
\name{get_final_projection}
\alias{get_final_projection}
\title{get final projections from simulation standings}
\usage{
get_final_projection(sim_standings, mode)
}
\arguments{
\item{mode}{}
}
\description{
get final projections from simulation standings
}
