% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_odds.R
\name{get_match_odds}
\alias{get_match_odds}
\title{get match odds from simmed results}
\usage{
get_match_odds(simmed_results, trials)
}
\arguments{
\item{trials}{}
}
\description{
get match odds from simmed results
}
