% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projection_odds.R
\name{get_projection_odds}
\alias{get_projection_odds}
\title{get projection_odds}
\usage{
get_projection_odds(final_projections, sim_standings, europe = 6, rel = 18)
}
\arguments{
\item{rel}{}
}
\description{
get projection_odds
}
