% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches_to_fixtures.R
\name{matches_to_fixtures}
\alias{matches_to_fixtures}
\title{convert list of matches to list of fixtures to be played}
\usage{
matches_to_fixtures(matches, date = Sys.Date())
}
\arguments{
\item{date}{}
}
\description{
convert list of matches to list of fixtures to be played
}
