% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches_to_team_ratings.R
\name{matches_to_team_ratings}
\alias{matches_to_team_ratings}
\title{convert matches to team ratings based on xG factor of played games}
\usage{
matches_to_team_ratings(matches, date = Sys.Date(), xG_factor = 0.6)
}
\arguments{
\item{date}{}
}
\description{
convert matches to team ratings based on xG factor of played games
}
