% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_to_standings.R
\name{results_to_standings}
\alias{results_to_standings}
\title{convert list of results to current standings}
\usage{
results_to_standings(results, xTable = F)
}
\arguments{
\item{xTable}{}
}
\description{
convert list of results to current standings
}
