% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_season_trials.R
\name{run_season_trials}
\alias{run_season_trials}
\title{run simulation of n random seasons of given remaining fixtures}
\usage{
run_season_trials(
  trials,
  fixtures,
  results,
  team_ratings,
  league_adj,
  xG_factor,
  smooth,
  hot
)
}
\arguments{
\item{results}{}
}
\description{
run simulation of n random seasons of given remaining fixtures
}
