% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_team_ratings.R
\name{update_team_ratings}
\alias{update_team_ratings}
\title{Title}
\usage{
update_team_ratings(current_team_ratings, gw_results, xG_factor, smooth, hot)
}
\arguments{
\item{hot}{}
}
\description{
Title
}
